+++ include/bertrand/bertrand.hpp
@@ -1,5 +1,5 @@
 /**
- * This file is part of "bertrand" version 0.0.14
+ * This file is part of "bertrand" version 0.0.15
  * https://github.com/bernedom/bertrand
  * A minimalistic, header only implementation of design by contract for C++
  *
@@ -52,8 +52,15 @@ static_assert(false, "Cannot enable and disable stacktrace at the same time");
 #include <regex>
 #include <string>
 #include <unistd.h>
+#else
+#undef BERTRAND_ENABLE_STACKTRACE
+#endif
+#endif
 
-void print_stacktrace(std::stringstream &output) {
+inline namespace bertrand {
+
+#if defined(BERTRAND_ENABLE_STACKTRACE)
+inline void print_stacktrace(std::stringstream &output) {
   output << "Stack trace:\n";
 
   // storage array for stack trace address data
@@ -115,12 +122,8 @@ void print_stacktrace(std::stringstream &output) {
   delete[] demangled_name;
   delete[] symbollist;
 }
-#else
-#undef BERTRAND_ENABLE_STACKTRACE
-#endif
-#endif
 
-namespace bertrand {
+#endif
 
 /// struct to check if value is contained in a list of values. i.e. to check
 /// if a value is valid for an enum